reorganize project structure, new util functions for KV, refactors
