+++ packages/lib/edge-cloudflare/example/wrangler.toml.example.toml
@@ -1,10 +1,15 @@
-name = "cloudflare"
-main = "src/index.ts"
-compatibility_date = "2024-03-14"
+name = "GrowthBook example edge app for CloudFlare"
+main = "index.ts"
+compatibility_date = "2024-05-07"
 
 # Variable bindings. These are arbitrary, plaintext strings (similar to environment variables)
 # Note: Use secrets to store sensitive data.
 # Docs: https://developers.cloudflare.com/workers/platform/environment-variables
+
+kv_namespaces = [
+  { binding = "KV_GB_CACHE", id = "abcde12345" },
+  { binding = "KV_GB_PAYLOAD", id = "qwerty123" }
+]
 
 [vars]
 NODE_ENV="production"
