+++ packages/lib/edge-utils/package.json
@@ -23,7 +23,6 @@
     "node-html-parser": "^6.1.13"
   },
   "devDependencies": {
-    "@types/jsdom": "^21.1.6",
     "@types/node": "^20.8.2",
     "rimraf": "^5.0.5",
     "typescript": "5.2.2"
