+++ packages/lib/edge-utils/tsconfig.json
@@ -5,6 +5,11 @@
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "target": "es6",
+    "lib": [
+      "es6",
+      "dom",
+      "es2016"
+    ],
     "noImplicitAny": true,
     "moduleResolution": "node",
     "sourceMap": true,
